+++ MEME-DOMAIN/src/main/java/org/meme/domain/entity/Inquiry.java
@@ -28,12 +28,4 @@ public class Inquiry extends BaseEntity {
     @JoinColumn(name = "user_id", nullable = false)
     private User user;
 
-//    public static Inquiry from(MypageInquiryDto dto, User user) {
-//        return Inquiry.builder()
-//                .inquiryText(dto.getInquiryText())
-//                .inquiryTitle(dto.getInquiryTitle())
-//                .user(user)
-//                .build();
-//    }
-
 }
