+++ MEME-SERVICE/src/main/java/org/meme/service/controller/ArtistController.java
@@ -4,7 +4,6 @@ import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.meme.domain.common.BaseResponseDto;
 import org.meme.domain.common.status.SuccessStatus;
-import org.meme.service.dto.ArtistRequest;
 import org.meme.service.service.ArtistService;
 import org.springframework.web.bind.annotation.*;
