+++ MEME-SERVICE/src/main/java/org/meme/service/controller/MypageController.java
@@ -4,8 +4,7 @@ import io.swagger.v3.oas.annotations.Operation;
 import lombok.RequiredArgsConstructor;
 import org.meme.domain.common.BaseResponseDto;
 import org.meme.domain.common.status.SuccessStatus;
-import org.meme.service.dto.ArtistRequest;
-import org.meme.service.dto.MypageRequest;
+import org.meme.service.dto.request.MypageRequest;
 import org.meme.service.service.MypageService;
 import org.springframework.web.bind.annotation.*;
