+++ MEME-SERVICE/src/main/java/org/meme/service/dto/response/ReviewResponse.java
@@ -1,4 +1,4 @@
-package org.meme.service.dto;
+package org.meme.service.dto.response;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
