+++ MEME-SERVICE/src/main/java/org/meme/service/service/ArtistService.java
@@ -9,8 +9,7 @@ import org.meme.domain.repository.ArtistRepository;
 import org.meme.domain.repository.FavoriteArtistRepository;
 import org.meme.domain.repository.ModelRepository;
 import org.meme.service.converter.ArtistConverter;
-import org.meme.service.dto.ArtistRequest;
-import org.meme.service.dto.ArtistResponse;
+import org.meme.service.dto.response.ArtistResponse;
 import org.springframework.stereotype.Service;
 import org.meme.domain.entity.Artist;
 import org.meme.domain.common.exception.GeneralException;
