+++ MEME-SERVICE/src/main/java/org/meme/service/service/FavoriteService.java
@@ -8,9 +8,9 @@ import org.meme.domain.entity.*;
 import org.meme.domain.repository.*;
 import org.meme.service.converter.ArtistConverter;
 import org.meme.service.converter.FavoriteConverter;
-import org.meme.service.dto.ArtistResponse;
-import org.meme.service.dto.FavoriteRequest;
-import org.meme.service.dto.FavoriteResponse;
+import org.meme.service.dto.response.ArtistResponse;
+import org.meme.service.dto.request.FavoriteRequest;
+import org.meme.service.dto.response.FavoriteResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
