+++ MEME-SERVICE/src/main/java/org/meme/service/service/PortfolioService.java
@@ -11,8 +11,8 @@ import org.meme.domain.entity.FavoritePortfolio;
 import org.meme.domain.repository.FavoritePortfolioRepository;
 import org.meme.domain.entity.Model;
 import org.meme.domain.repository.ModelRepository;
-import org.meme.service.dto.PortfolioRequest.*;
-import org.meme.service.dto.PortfolioResponse.*;
+import org.meme.service.dto.request.PortfolioRequest.*;
+import org.meme.service.dto.response.PortfolioResponse.*;
 import org.meme.domain.entity.Portfolio;
 import org.meme.domain.entity.PortfolioImg;
 import org.meme.domain.repository.PortfolioImgRepository;
