+++ MEME-SERVICE/src/main/java/org/meme/service/service/ReviewService.java
@@ -7,8 +7,8 @@ import org.meme.domain.entity.*;
 import org.meme.domain.enums.Status;
 import org.meme.domain.repository.*;
 import org.meme.service.converter.ReviewConverter;
-import org.meme.service.dto.ReviewRequest;
-import org.meme.service.dto.ReviewResponse;
+import org.meme.service.dto.request.ReviewRequest;
+import org.meme.service.dto.response.ReviewResponse;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.PageRequest;
